Edit Wrapping IMU class(Supports two IMUs with one code)
