+++ examples/Basics/IMU/IMU.ino
@@ -8,40 +8,51 @@ float gyroX = 0.0F;
 float gyroY = 0.0F;
 float gyroZ = 0.0F;
 
+float pitch = 0.0F;
+float roll  = 0.0F;
+float yaw   = 0.0F;
+
 void setup() {
   // put your setup code here, to run once:
   M5.begin();
-  M5.Sh200Q.Init();
+  M5.IMU.Init();
   M5.Lcd.setRotation(3);
   M5.Lcd.fillScreen(BLACK);
   M5.Lcd.setTextSize(1);
   M5.Lcd.setCursor(40, 0);
-  M5.Lcd.println("SH200I TEST");
-  M5.Lcd.setCursor(0, 15);
+  M5.Lcd.println("IMU TEST");
+  M5.Lcd.setCursor(0, 10);
   M5.Lcd.println("  X       Y       Z");
+  M5.Lcd.setCursor(0, 50);
+  M5.Lcd.println("  Pitch   Roll    Yaw");
 }
 
 float temp = 0;
 /*****************************************
-M5.Sh200Q.getGyroData(&gyroX,&gyroY,&gyroZ);
-M5.Sh200Q.getAccelData(&accX,&accY,&accZ);
-M5.Sh200Q.getTempData(&temp);
+M5.IMU.getGyroData(&gyroX,&gyroY,&gyroZ);
+M5.IMU.getAccelData(&accX,&accY,&accZ);
+M5.IMU.getAhrsData(&pitch,&roll,&yaw);
+M5.IMU.getTempData(&temp);
 *****************************************/
 void loop() {
   // put your main code here, to run repeatedly:
-  M5.Sh200Q.getGyroData(&gyroX,&gyroY,&gyroZ);
-  M5.Sh200Q.getAccelData(&accX,&accY,&accZ);
-  M5.Sh200Q.getTempData(&temp);
+  M5.IMU.getGyroData(&gyroX,&gyroY,&gyroZ);
+  M5.IMU.getAccelData(&accX,&accY,&accZ);
+  M5.IMU.getAhrsData(&pitch,&roll,&yaw);
+  M5.IMU.getTempData(&temp);
   
-  M5.Lcd.setCursor(0, 30);
+  M5.Lcd.setCursor(0, 20);
   M5.Lcd.printf("%6.2f  %6.2f  %6.2f      ", gyroX, gyroY, gyroZ);
-  M5.Lcd.setCursor(140, 30);
+  M5.Lcd.setCursor(140, 20);
   M5.Lcd.print("o/s");
-  M5.Lcd.setCursor(0, 45);
+  M5.Lcd.setCursor(0, 30);
   M5.Lcd.printf(" %5.2f   %5.2f   %5.2f   ", accX, accY, accZ);
-  M5.Lcd.setCursor(140, 45);
+  M5.Lcd.setCursor(140, 30);
   M5.Lcd.print("G");
   M5.Lcd.setCursor(0, 60);
+  M5.Lcd.printf(" %5.2f   %5.2f   %5.2f   ", pitch, roll, yaw);
+
+  M5.Lcd.setCursor(0, 70);
   M5.Lcd.printf("Temperature : %.2f C", temp);
   delay(1000);
 }
